Add PersistentSurveysSegmentsAdapter
This is not completely implemented yet so inherits most behaviour
from the session-based SurveysSegmentsAdapter.

This will be confusing in the future so we should try to avoid
using the session-based adapter entirely.

However we need to get this into production reasonably quickly to
allow us to start storing this data.

I'm using `apps.get_model()` to prevent a circular import.