Fetch tag count data with the persistent segment adapter
This commit uses the persisted data rather than the session data
to calculate whether a user should be in a segment based on what
tagged articles they've visited.

This will cause one database query per pageview for a logged-in user,
per segment that is currently enabled. The query is of the form:

SELECT COUNT(*)
    FROM (SELECT page_id, COUNT(page_id) FROM "table"
        WHERE (visited_at >= $TIMESTAMP AND visited_at <= $TIMESTAMP AND tag_id = $ID AND user_id = $ID)
        GROUP BY page_id)

This isn't terrible because at least it's only using one table, but it's
still not ideal.